# server/swagger.yaml
openapi: 3.0.0
info:
  title: Kanata Scraper API
  description: API documentation for Kanata Scraper (Otakudesu, etc.)
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local Development Server
paths:
...
  /api/neonime/search:
    get:
      tags:
        - Neonime
      summary: Search Anime on Neonime
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema: { type: array, items: { type: object } }
  /api/neonime/anime/{slug}:
    get:
      tags:
        - Neonime
      summary: Get Neonime Anime Details
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Anime details
          content:
            application/json:
              schema: { type: object }
...